# src/swagger/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OHC账票生成服务",
    "description": "OHC账票生成API服务",
    "version": "1.0.0"
  },
  "paths": {
    "/templates": {
      "get": {
        "tags": [
          "templates"
        ],
        "summary": "获取模板列表",
        "description": "获取所有支持的模板列表",
        "operationId": "list_templates_templates_get",
        "parameters": [
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Language"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/templates/{template_name}": {
      "get": {
        "tags": [
          "templates"
        ],
        "summary": "获取模板信息",
        "description": "获取指定模板的详细信息",
        "operationId": "get_template_info_route_templates__template_name__get",
        "parameters": [
          {
            "name": "template_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Template Name"
            }
          },
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Language"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateInfoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TemplateInfoResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "模板名称"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name",
            "description": "显示名称"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "模板描述"
          },
          "available_formats": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Available Formats",
            "description": "可用格式"
          },
          "filler_strategy": {
            "type": "string",
            "title": "Filler Strategy",
            "description": "填充策略"
          },
          "features": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Features",
            "description": "特性列表"
          },
          "available_languages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Available Languages",
            "description": "可用语言列表"
          },
          "display_names": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Names",
            "description": "各语言的显示名称"
          }
        },
        "type": "object",
        "required": [
          "name",
          "display_name",
          "available_formats",
          "filler_strategy",
          "features"
        ],
        "title": "TemplateInfoResponse",
        "description": "模板信息响应模型"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}